+++ src/adapters/openclaw/mcp-tools.ts
